requirements.txt
appdirs==1.4.3
gitdb2==2.0.0
GitPython==2.1.3
packaging==16.8
pep8==1.7.0
pyparsing==2.2.0
PyYAML==3.12
six==1.10.0
smmap2==2.0.1
